Don't enforce MaxItems early in the validation process
Having multiple dynamic blocks for the same block, where only one of them will actually resolve to a real block, doesn't pass validation because Terraform sees both blocks and thinks MaxItems is violated.  This changes things (similar to how #22221 did for MinItems) so that the validation is relaxed until later on when the dynamic blocks are resolved.